Created and passed basic tests for debit card getters

+++ src/main/java/CreditCard.java
@@ -3,7 +3,7 @@ public class CreditCard extends PaymentCard {
     private double riskMultiplier;
     private double creditLimit;
 
-    public CreditCard(int cardNumber, String expiryDate, int securityNumber, String provider, double riskMultiplier, double creditLimit) {
+    public CreditCard(long cardNumber, String expiryDate, int securityNumber, String provider, double riskMultiplier, double creditLimit) {
         super(cardNumber, expiryDate, securityNumber, provider);
         this.riskMultiplier = riskMultiplier;
         this.creditLimit = creditLimit;
